refactor: outbox 전환에 따른 kafka producer 설정 테스트로 이동
